REPLACNONES: Broken
added a config option that determines wheter values which are equal to None,
or of NoneType, should be replaced with an empty string "".

This is usefull, as some items will be reported as "unsupported item" once
the value "None" is received by the zabbix server.

Imho using the empty string is the best option to deal with this instead of
filtering out the values completely, because None can be a valid option, for
instance for wind directions, when no wind is present.

Sending an empty string should circumvent this issue. but zabbix sender fails
with an error: 'ERROR: [line 26] 'Key value' required Sending failed. Use
option -vv for more detailed output.'

Thus this commit breaks the application.